svg/animations/repeat-iteration-event-004.svg

<svg xmlns="http://www.w3.org/2000/svg"><title>Animation triggers on &apos;repeat(3)&apos;</title><script xmlns="http://www.w3.org/1999/xhtml" src="/resources/testharness.js"/><script xmlns="http://www.w3.org/1999/xhtml" src="/resources/testharnessreport.js"/><rect width="0" height="100" fill="green"><animate id="anim" attributeName="visibility" begin="0s" dur="2s" repeatCount="4" to="visible"/><set attributeName="width" begin="anim.repeat(3)" to="100"/></rect><script>promise_test(t =&gt; { document.documentElement.setCurrentTime(5.995); let set = document.querySelector(&quot;set&quot;); return new EventWatcher(t, set, &apos;beginEvent&apos;).wait_for(&apos;beginEvent&apos;).then(() =&gt; { assert_equals(set.parentNode.getBBox().width, 100); }); });</script></svg>
